feat(grid): add image grid component
